task 5 transposeIfMatrixHasNotEqualSumOfRows

+++ libs/data_structures/matrix/matrix.h
@@ -99,7 +99,7 @@ matrix mulMatrices(matrix m1, matrix m2);
 bool isUnique(long long *a, int n);
 
 //возвращает сумму элементов массива.
-long long getSum(int *a, int n);
+int getSum(int *a, int n);
 
 //проводит автоматизированное тестирование библиотеки
 void test();
